Refactor FeaturedWork component to use Suspense and create FeaturedWorkClient for improved loading and animation
